Move serializer files to subfolder
