+++ app/stac_api/serializers/serializers_utils.py
