+++ app/stac_api/views/views.py
@@ -33,13 +33,13 @@ from stac_api.models import LandingPage
 from stac_api.pagination import ExtApiPagination
 from stac_api.pagination import GetPostCursorPagination
 from stac_api.s3_multipart_upload import MultipartUpload
-from stac_api.serializers import AssetUploadPartsSerializer
-from stac_api.serializers import AssetUploadSerializer
-from stac_api.serializers import CollectionAssetUploadSerializer
-from stac_api.serializers import ConformancePageSerializer
-from stac_api.serializers import ItemSerializer
-from stac_api.serializers import LandingPageSerializer
-from stac_api.serializers_utils import get_relation_links
+from stac_api.serializers.serializers import AssetUploadPartsSerializer
+from stac_api.serializers.serializers import AssetUploadSerializer
+from stac_api.serializers.serializers import CollectionAssetUploadSerializer
+from stac_api.serializers.serializers import ConformancePageSerializer
+from stac_api.serializers.serializers import ItemSerializer
+from stac_api.serializers.serializers import LandingPageSerializer
+from stac_api.serializers.serializers_utils import get_relation_links
 from stac_api.utils import call_calculate_extent
 from stac_api.utils import get_asset_path
 from stac_api.utils import get_collection_asset_path
