+++ app/tests/tests_10/test_serializer.py
@@ -13,9 +13,9 @@ from rest_framework.renderers import JSONRenderer
 from rest_framework.test import APIRequestFactory
 
 from stac_api.models import get_asset_path
-from stac_api.serializers import AssetSerializer
-from stac_api.serializers import CollectionSerializer
-from stac_api.serializers import ItemSerializer
+from stac_api.serializers.serializers import AssetSerializer
+from stac_api.serializers.serializers import CollectionSerializer
+from stac_api.serializers.serializers import ItemSerializer
 from stac_api.utils import get_link
 from stac_api.utils import isoformat
 from stac_api.utils import utc_aware
