+++ app/tests/tests_10/test_serializer_asset_upload.py
@@ -7,7 +7,7 @@ from rest_framework import serializers
 from rest_framework.exceptions import ValidationError
 
 from stac_api.models import AssetUpload
-from stac_api.serializers import AssetUploadSerializer
+from stac_api.serializers.serializers import AssetUploadSerializer
 from stac_api.utils import get_sha256_multihash
 
 from tests.tests_10.base_test import StacBaseTestCase
